+++ new-app/src/Profile.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Profile() {
+  return (
+      
+    <h1>Profile</h1>
+  );
+}
+export default Profile;
